+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 ## [0.0.1] - Apr 28 2020.
 
-* TODO: Initial release.
+* Initial release.
 
 ## [0.0.2] - Jun 08 2020.
 
 * Fixed Crash on **onCountDownComplete**.
+* Updated **README**.
+
+## [0.0.3] - Jun 13 2020.
+
+* Changed Parameters name from:
+    * **countdownTextStyle** to **textStyle**.
+    * **reverseOrder** to **isReverse**.
+    * **onCountDownComplete** to **onComplete**.
+
 * Updated **README**.